+++ src/main/java/com/example/demo/dto/DisciplinaDTOResponse.java
@@ -12,5 +12,5 @@ public class DisciplinaDTOResponse {
     private Long id;
     private String nome;
     private Boolean active;
-    private Byte media;
+    private Float media;
 }
